rest/src/requirements.txt
pyramid ==1.10
mysql-connector-python
requests==2.18